pyproject.toml
[project]
name = "ergo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dspy>=2.6.27",
    "nltk>=3.9.1",
    "psycopg2>=2.9.10",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "sentence-transformers>=4.1.0",
]
